KD-51 - Acquisition integration - Creates a basis for multivendor acquisitions integration to Koha
Displays a new feature, "Send the order" in basketgroups->closed -view. This sends the basketgroup to
any other vendor using the new AcquisitionIntegration-module, if an interface is defined.

Introduces a new state for a basketgroup, ordered. Which is set by a (deliverycomment =~ /tilattu/).
This displays a small notification in basketgroups-view if an order is already ordered.

Displays the deliverycomment-column in basketgroups->closed -view to better understand the ordered-state.

Disables printing a pdf when closing a basketgroup.

Introduces pullSelectionListsFromBookvendors.pl -cronjob to daily stage selection lists to Koha and to stage
BTJ full biblios to kohazebra.