KD-122 - Checkin Items' holdingbranch to transfer destination fix
Checkin an item which is checked out, to a non-homebranch.
Item's holdingbranch is set to the homebranch, even if it should be the place of checkin.
This is because a transfer is initiated ( ModItemTrans() ) and this function for some reason sets the holdingbranch to the transfer destination.
On the second checkin to the same branch this doesn't happen, because transfer is already initialized and there is no checkout active, thus ModItemTrans() is not triggered.

Most definetely this is a bug, but unsure what is the proper way to fix this.
Solving this using the same method from C4/Circulation->updateWrongTransfer() where the holdingbranch is simply reverted to what it should be.