KD-130: Usability improvements to the cataloguing/additem -view and the acqui/neworderempty -view
This feature populates the itemcallnumber field (952$o) from the Biblio's main heading, call number
and from the OPLIB Labels generator.

Values are rechecked every time the homebranch, itemtype, callnumber or the shelving location changes,
or when the itemcallnumber is manually changed/emptied.

Only the shelving code is changed automatically and other itemcallnumber components need to be
manually changed or the whole itemcallnumber can be rebuilt.

Refactored modules to improve clarity and to work with KD-204.