Bug 11175 - (KD-16) Displaying component parts in OPAC and SC, Plumbing
This commit creates the basis for using component parts in the detailed views.
A full Zebra search index rebuilding is required for the new field 003 to be indexed.

TERMINOLOGY:
"Component part": A Bibliographic record connected to it's host record via $773w -> 001
  or via other Record-control-number-index relation.
"Host record": A bibliographic record which contains child records. Host record doesn't
  need to have a link to the child record.

USEFUL:
For music cataloguers whose jobs depend on these.
For browsing the contents of a musical record.
Provides the basis for a high quality music library service.

Because Koha has the building blocks to create these component part records out of box, it is odd
that they can't be displayed out of box.

This patch adds a system preference AddComponentPartRecordsToDetailedViews.
This controls if component part records data is appended after the bibliographic data for XSLT parsing.
Also makes sure that component part records are not pulled needlessly for non-detailed views as
this would cripple performance.

WHAT DID I DO?!
-Added indexing definitions to NORMARC and MARC21, for field 003, Control-number-identifier, cni.
-Added Control-number-identifier and cni to acceptable search keywords hashes, so they can be parsed
  as proper CCL/PQF queries.
-Added a system preference AddComponentPartRecordsToDetailedViews to toggle the display of component parts
  in detailed views (component parts are NEVER parsed for the result view due to performance issues!)
-Added accessors to C4::Biblio to get component parts easily.