Bug 12624 - Getting all search results in C4::Breeding::Z3950Search, instead of pageful
Fetching ALL component parts from a remote Z3950 repository is difficult.
Also the Z3950Search() function only works directly to templates.

This patch changes the Z3950Search() to OPTIONALLY return all search results as a HASH instead of a Template-object.
Previous functionality is left untouched.

This makes scripting with it a LOT easier!