Bug 11677 - Limit to Only items currently available for loan or reference not working!
This row:
$availability_limit .=
"( ( allrecords,AlwaysMatches='' not onloan,AlwaysMatches='') and (lost,st-numeric=0) )";
Causes all records in Zebra to be selected and removes records with even one item onloan
  (even if items would be available)!
This is obviously wrong, but because Zebra (and not many other text indexers) cannot deal
with (If undef)-search terms, a fix might be to set 952$q (onloan) datetime to 0000-00-00
instead of NULL in the DB. Then we could find Records with items where onloan is 0000-00-00.

Comments out that Zebra availability search and deals with availability on Koha-side.

Gets the branch or holdingbranch limit from opac-search.pl or search.pl and uses that to
skip Zebra results which are onloan or notforloan.
Updates the result set size when results are removed.

Fixed a bug where pagination broke, due to the $results_hash->{'RECORDS'} -array getting populated
always starting from 0, where it should have started from the given $offset.