Cronjob monitoring script, nagios3 compatible
After having issues with cronjobs not always running or getting regression issues, I decided to give a try
to create a cronjobs monitoring solution.

This patch introduces a wrapper script to log and time cronjobs.

The wrapper script's log output is used to analyze whther cronjobs ran and if there were any errors.
Also cronjobs timings can be validated to prevent nasty errors with for. ex. expiring waiting holds one day too
early/late.
This check can also be integrated to nagios!

Also provided is a scaffold to start introducing more nagios-checks, with example command and service
configurations.

This patch introduces a koha_cronjobs-script to read the /var/spool/cron/crontabs/koha and analysz its contents.
Then one can validate whether the cronjob timings are right (this is not mentioned in the manual at all) and what
is right is (currently) based on my honest feelings.
Then each cronjobs frequency is determined and the cronjobs log output is checked based on the script parameters.
Currently the cronjob frequencies are split to minutely, hourly, daily, etc.
Koha_cronjobs-script is meant to be ran on those intervals to catch errors ASAP.