Bug 1985 - Email notification of new OPAC comments
This feature adds YAS 'CommentModeratorsEmail', where an email (template code COMMENT_CREATED) is sent when
a borrower creates/modifies an comment/review. The syspref is empty by default and no email is sent when the
syspref is empty.

This feature is needed because we want to alert our content moderators to take action when needed. Usually
it is our librarians who write reviews and this improvement makes the process of moving Koha reviews to our
CMS more smooth.

TEST SETUP:
-1. Run updatedatabase.pl to add the syspref and the new letter template.
0. Go to admin/preferences.pl?tab=opac, and set the 'CommentModeratorsEmail'-syspref to your email-address.

TEST PLAN:
1. Log in to OPAC and find any Biblio.
2. Make a comment/review for it.
3. Check the koha.message_queue-table for any letter_code = 'COMMENT_CREATED'. You should see the reasonable
   default message. You can change the message in tools/letter.pl.