Bug 13471 - When editing a subscription, the "subscription length type" disregards the stored value, also make SLT translateable.
This issue causes loss of data when the subscription length type (SLT) is other than issues.

TO REPLICATE:

We make an subscription with SLT of months = 12.
Later make an arbitrary edit to the subscription, and miss out the fact that the SLT selection is set to issues.
Save the changes.
Serial subscription end date is set to 12 issues since the start date, instead of 12 months which it used to be.
This causes a myriad of issues :=)

AFTER THIS PATCH:

SLT defaults to the value initially saved instead of always being issues.
Now the selection can be translated!

HOTFIX 1: Added <span> elements to showSubscriptionLengthType.inc so the translation tool can catch them.

HOTFIX 2: Fixed serials/showpredictionpattern.pl to use the db columns instead of issues,weeks,months.

HOTFIX 3: Fixed serials/subscription-numberpatterns.pl to use the db columns instead of issues,weeks,months.