Bug 13540 - Item's permanent_location is set to CART|PROC if an Item is edited when it's location is CART|PROC.
This bug is rather nasty as it happens very discreetly when editing Items that are
in either locations 'CART' or 'PROC'.
For example when changing one of the Item's statuses or barcode, the permanent_location
gets overwritten without the user noticing it immediately. The damage becomes apparent
only weeks later when there are potentially thousands of Items missing the
permanent_location and location.

:-:TEST SETUP:-:
0a. System preference "ReturnToShelvingCart" must be 'Move'.
0b. Find any Item (I69)
0c. Set I69's location to "Book Cart" (BC) or anything but empty.

:-:REPLICATE ISSUE:-:
1. Check-in the I69 and the location should turn to 'CART' (returned today).
2. Go to the "Edit Item"-view for I69, don't change anything, simply click
   "Save changes"
3. Observe that the permanent_location has changed to 'CART'.
   This is not good since now the location data is forever lost again (Bugg 7817).
   The "cart_to_shelf.pl --hours 24"-script cannot return the Item to the proper
   place ;(

:-:AFTER THIS PATCH:-:
Repeat steps 1-2.
3. Observe that the permanent_location is still the old one, instead of getting
   overwritten by 'CART'.