Bug 13707 - Adding a new column datereceived for items and biblioitems
This column tells when the Item or Biblio has been received for the first time
into a library.
Date-of-acquisition (dateaccessioned) answers, when the Item has been ordered.
Date-of-receival (datereceived) answers, when the Item has arrived to the
library or when the first Item for the Biblio has for the first time arrived to
the library.
This is very useful regarding statistics of acquisitions if they are based on
the date received instead of the date acquired.
Also other cool stuff can be built on top of it, like better searches for new
material.

This patch adds new DB columns items.datereceived, biblioitems.datereceived,
deleteditems.datereceived, deletedbiblioitems.datereceived
and populates those columns from the acorders.datereceived if the Item has been
acquired using the acquisitions module, or from the
items.dateaccessioned-column if no acquisitions module has been used.
Biblioitems.datereceived is always the oldest timestamp from all the Items the
Biblio has.