KD-245 - Bug 13707 - 6. notforloan-searchindex added to GUI for staff.
This is needed so staff can make more complex searches regarding the
datereceived-index.
Our acquisitions have "On order" status even after they have been received,
so we need to be able to search for Items without the "On order"-status to
get the available received items.

Added the year-limit for datereceived to limit to only this years, or past
years received Biblios.