Bug 13995 - Proper Exception handling
Koha has issues with propagating errors and exceptions from the business layer
to the front-end. Using a more sophisticated system of named Exceptions we can
send more formal signals towards the user.
Catching different types of Exceptions makes for a more concise handling of
errors.
Also throwing/catching Exceptions is a industry standard and for a good reason.
There is no point for Koha to lurk behind.

USAGE:
try {
    Koha::Exception::BadSystemPreference->throw(error => 'Syspref DisplayIconsXSLT is not a valid boolean');
} catch {
    if (blessed($_) && $_->isa('Koha::Exception::BadSystemPreference')) {
        print $_->as_string();
        warn $_->error, "\n", $_->trace->as_string, "\n";
    }
    else {
        $_->rethrow();
    }
}