SIP Server bugfix rescues a broken DB connection for the SIP worker.
After a long period of inactivity the SIP worker thread loses the connection
to the DB and regardless of the $dbh->{mysql_auto_reconnect} -parameter,
it cannot renew the connection.
This patch checks for dead connections every time a new SIP message is received,
and creates a brand new one if needed.