Buuug 14287 - Preserve valid ISBN-13 identifiers. REPLACE FROM master/bugzilla WHEN MIGRATING!!
If Business::ISBN cannot make an ISBN-10 out of a valid ISBN-13 starting with
979, then we should preserve the ISBN-13 instead of getting rid of it.
ISBN-13 starting with 979 is valid as well and in a few years starting with 980
might be as well as the ISBN supply is running out.

This patch modifies the C4::Koha::NormalizeISBN() to preserve the original ISBN
if type conversion fails.

Use Buug 13167's TEST PLAN.