KD-38, KD-37. Claiming. Fix to include all accountlines for a claimed Item.
If an Item has incurred fines and is then renewed, then the Item starts
collecting new fines in a different accountlines-entry.
Claiming module only picks one accountline entry to use as the incurred fine.

This commit fixes to use the SUM of all incurred fines for the Item for the borrower.