Bug 14680 - when doing acquisitions from a staged file, MarcFieldsToOrder-syspref discounts are instead added.
Using the syspref 'MarcFieldsToOrder', getting the discount directly from the
MARC causes the price to actually increase by the discount percentage, not
decrease, if the Vendor/Bookseller is configured to include the VAT/GST in the
list price.

REPLICATE ISSUE:

-1. Define 'MarcFieldsToOrder'-syspref to match the discount and price columns
    of your staged MARC Records.
0. Stage a MARC batch.
1. Make and acquisition from the staged file using
       /acqui/addorderiso2709.pl (Save the big staging list, don't click
                                  "Add order" on an individual Biblio.)
2. Observe that the Replacement price (rrp) is actually the (price + discount)
   instead of (price - discount).
   And the estimated cost is just the price in the MARC Record without any
   discount.

AFTER THIS PATCH:

Replay step 1.
2. rrp is now the price in the MARC Record incl. VAT/GST.
   ecost is now the VAT/GST included discounted price.

NOTE!

Because the subroutine I modified doesn't make much sense, it is presumed that
there can be regression in other parts for other kinds of vendor configurations.
Because the acquisitions module is in urgent need of complete back-end
refactoring + PageObject/Cucumber regression test coverage, no further
investigation is made to resolve those.