Bug 14767: (follow-up) Use Unicode-parameter only when needed
This patch makes Labyrintti driver use Unicode parameter only when required.
In most cases in our local library we are fine with GSM 03.38. We will check
if data is lost when comparing the decoded gsm0338 encoded message to the
original message. Add it as optional feature in koha-conf - smsProviders -
labyrintti - Unicode. To turn it on, set the value to "yes".

For statistics, store the result in "metadata" field of message_queue
if the message was sent as UTF-16.

We also make a small fix to the response protocol since it is not in sync with
what Labyrintti expects. Their delivery report wants a 200 OK response, not 201
or 204.