Bug 14872: Reminder to update SMSnumber when primary phone is updated
Since SMS number (smsalertnumber) and primary phone are separated, it would
be useful to remind patron to update his/her SMS number when making changes
to primary phone. This can easily be forgotten, especially in OPAC where the
two fields are in different page (opac-memberentry.pl vs opac-messaging.pl).

This patch adds a reminder in the "hint"-div that activates whenever phone
field gets input (and phone-smsalertnumber do not match).

To test:
1. Enable SMSSendDriver and EnhancedMessagingPreferences system preferences
2. Navigate to OPAC -> your personal details
3. Input something to primary phone. You should not be notified at all
4. Navigate to Intra -> Patron -> Edit
5. Repeat step 3
6. Apply patch
7. Refresh Opac -> your personal details and Intra -> Patron -> Edit
8. Input something to primary phone and observe you will be notified to
   also update SMS number.