KD#377 - CPU / Ceepos integration. Integrate cash registers (POS) into Koha.
This commit integrates Ceepos cash registers into Koha. The basic idea is to
construct a payment from Koha accountlines and send it to the cash register
where it will be either accepted or cancelled. CPU will notify Koha the payment
status. After this, Koha modifies the accountlines accordingly and creates a
new accountline of type "Pay".

Changes in views for staff client are kept minimal. This patch does not
include OPAC payments and has nothing to do with OPAC at all.

Key features are:
  - Create Koha::Object for transactions called "PaymentTransactions".
  - Stop using C4::Accounts default functions because they suck. Instead, create
    a new subroutine that combines all four C4::Accounts way of payment. The new
    subroutine "completes" a transaction, i.e. after payment is successful, it
    makes an old-style Koha-payment so that Koha is in sync with the cash register
    system. Call by $transaction->CompletePayment();
  - members/paycollect.pl creates a new transaction and sends it via Ajax.
    (C4::OPLIB::CPUIntegration::SendPayment() sends it)
  - REST endpoint for getting a transaction:
      - /api/v1/pos/cpu/{invoicenumber} GET
  - REST endpoint for receiving reports from CPU:
      - /api/v1/pos/cpu/{invoicenumber}/report POST
  - System preferences
      - cpuitemnumbers: YAML config that maps Koha accountlines itemtypes to actual
                        cash register codes.
      - POSIntegration: turn the whole integration on/off.
  - Does not use the integration for branches that have not mapped Koha-items to
    CPU items in "cpuitemnumbers"
Configurations to get started:
  - koha-conf.xml introduces new settings for CPU (pos->CPU):
      - source: Source key provided by CPU
      - secretKey: Secret key to construct a SHA-256 hash from input
      - url: address to CPU
      - modle: defines the mode for response type (1: synchronous, 2: asynchronous)
      - notificationAddress: address where we want to receive the payment report
      - also paths for ssl certificate (CPU requires this)
  - System preferences
     - mentioned in key features above. Set them up to get started.

Step-by-step description of how it all works:

Staff member enters Borrower's Fines and goes to "Pay fines" tab. For any fine in
the table, he clicks "Pay". He is then given an option to define the total amount
for the payment. THE WHOLE AMOUNT IN THE INPUT BOX MUST BE PAID AT CASH REGISTER!
He then selects a cash register (if there is none, he clicks "Add new cash register"
and enters the cash register code number). Then he clicks "Confirm".

Here's where the integration really starts to take action.

In members/paycollect.pl, we call C4::OPLIB::CPUIntegration::InitializePayment()
with the selected fines. If the staff member did not define any specific items,
InitializePayment() will choose which items will be paid. InitializePayment()
will create a new row into payments_transactions table. The items that belong
to this payment, are added into payments_transactions_accountlines with detailed
price for each item.

After that, a new view for the staff member is presented. This is the screen that
shows when the payment is pending. AJAX call is made to
/cgi-bin/koha/members/paycollect.pl. Then, Koha sends the payment to CPU from
C4::OPLIB::CPUIntegration::SendPayment().

When the payment is completed, CPU responds to Koha's long-polling call. If the call
is timed out, the completion of a payment is made when CPU sends a report to REST API.
The payment is completed by calling $transaction->CompletePayment("paid"). This subroutine
modifies amountoutstanding of all of the items in accountlines related to the payment.
It then creates a new line into accountlines with the type "Pay" and sets the amount
equal to the amount that was paid. We now have successful payment. The events are also
logged.

Payment can be also reverted via the cash register. In this case, we call
$transaction->RevertPayment() and it will revert the payment.

To install:
1. Apply patch
2. Run atomicupdater
3. Install SSL certificates delivered by CPU.
4. Edit koha-conf. Add parameters given by CPU and paths to SSL certificates.
5. Edit koha-httpd. Block access to NotificationAddress from others except CPU.
6. Restart koha-api-daemon.
7. Set system preferences
7.1 cpuitemnumbers: Map Koha itemtypes (accountlines.accounttype) into Ceepos itemnumbers
                    for each library. If not defined for library, integration is considered
                    offline in that library.
7.2 POSIntegration -> CPU integration

Also make sure there will be no conflicts with payment ids. Payment IDs are unique! Payment id
is the transaction_id in payments_transactions table.

To use:
1. Make sure Ulkop. liityntä is enabled in Ceepos.
2. Make sure all connections are open between Koha and CPU server.
3. Go to pay Patron's payments. Add a new Office (possibly equal to cash register number).
   This office will stay in browser's cookie until it is deleted.
4. Select the Office and pay the fine.
5. Click Ulkop. liityntä (or SHIFT+F7) in Ceepos
6. You should see the fine in Ceepos. Proceed with the cash register as before.
7. After payment is complete, Koha will be notified automatically. You don't have to worry
   about anything. Just make sure the payment is correctly marked into Borrower's account.

See KD#377 in Redmine for more detailed description.